+++ src/main/java/org/example/srg3springminiproject/service/serviceImpl/ExpenseServiceImpl.java
@@ -1,6 +1,7 @@
 package org.example.srg3springminiproject.service.serviceImpl;
 
 import lombok.AllArgsConstructor;
+import org.example.srg3springminiproject.exception.NotFoundException;
 import org.example.srg3springminiproject.model.Expense;
 
 import org.example.srg3springminiproject.model.request.ExpenseRequest;
@@ -9,6 +10,7 @@ import org.example.srg3springminiproject.service.ExpenseService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.UUID;
 
 @Service
 @AllArgsConstructor
@@ -16,44 +18,57 @@ public class ExpenseServiceImpl implements ExpenseService {
     private final ExpenseRepository expenseRepository;
     private final UserServiceImpl userService;
 
-
     @Override
-    public Expense findExpenseById(Integer id) {
-        Long userId = userService.getUsernameOfCurrentUser();
-        return expenseRepository.findExpenseById(id,userId);
-    }
+    public Expense findExpenseById(UUID id) {
+        UUID userId = userService.getUsernameOfCurrentUser();
 
+        Expense expense = expenseRepository.findExpenseById(id,userId);
+        //return expenseRepository.findExpenseById(id,userId);
+
+        if (expense == null) {
+            throw new NotFoundException("The expense with id " + id + " doesn't exist.");
+        }
+        else {
+            return expenseRepository.findExpenseById(id,userId);
+        }
+    }
 
     @Override
 
     public List<Expense> getAllExpense(int offset, int limit, String sortBy,String orderByStr) {
-        Long userId = userService.getUsernameOfCurrentUser();
+        UUID userId = userService.getUsernameOfCurrentUser();
         offset = (offset - 1) * limit;
         return expenseRepository.getAllExpense(offset,limit,sortBy,orderByStr, userId);
     }
 
-
     @Override
     public Expense saveExpense(ExpenseRequest expenseRequest) {
-        Long userId = userService.getUsernameOfCurrentUser();
+        UUID userId = userService.getUsernameOfCurrentUser();
         Expense expenseId = expenseRepository.saveExpense(expenseRequest,userId);
         return expenseId;
-
     }
 
-
     @Override
-    public Expense updateExpense(Integer id, ExpenseRequest expenseRequest) {
-        Long userId = userService.getUsernameOfCurrentUser();
+    public Expense updateExpense(UUID id, ExpenseRequest expenseRequest) {
+        UUID userId = userService.getUsernameOfCurrentUser();
         Expense expenseId = expenseRepository.updateExpense(id,expenseRequest,userId);
-        return expenseId;
-    }
+        //return expenseId;
 
+        if (expenseId == null) {
+            throw new NotFoundException("The expense with id " + id + " doesn't exist.");
+        }
+        else {
+            return expenseId;
+        }
+    }
 
     @Override
-    public Boolean deleteExpense(Integer id) {
-        return expenseRepository.deleteExpense(id);
+    public Boolean deleteExpense(UUID id) {
+        if (!expenseRepository.deleteExpense(id)) {
+            throw new NotFoundException("The expense with id " + id + " doesn't exist.");
+        }
+        else {
+            return expenseRepository.deleteExpense(id);
+        }
     }
-
-
 }
